# docs/swagger.yaml
basePath: /
definitions:
  controllers.InvoiceViewFormat:
    properties:
      invoice_id:
        type: string
      order_details: {}
      order_id:
        type: string
      payment_due: {}
      payment_due_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      table_number: {}
    type: object
  models.Food:
    properties:
      created_at:
        type: string
      food_id:
        type: string
      food_image:
        type: string
      id:
        type: string
      menu_id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    required:
    - food_image
    - menu_id
    - name
    - price
    type: object
  models.Invoice:
    properties:
      created_at:
        type: string
      id:
        type: string
      invoice_id:
        type: string
      order_id:
        type: string
      payment_due_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      updated_at:
        type: string
    required:
    - order_id
    - payment_method
    - payment_status
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: CRUP api's for go backend
  title: go-backend
  version: "1.0"
paths:
  /foods:
    get:
      description: Returns an array of foods from the ordercollection in restaurent
        database.
      produces:
      - application/json
      responses:
        "200":
          description: slice of orders
          schema:
            items:
              $ref: '#/definitions/models.Food'
            type: array
      security:
      - Security.require(true): []
      summary: Returns slice of foods
      tags:
      - foods
    post:
      consumes:
      - application/json
      description: Creates a food resource on the server
      parameters:
      - description: Food object
        in: body
        name: food
        required: true
        schema:
          $ref: '#/definitions/models.Food'
      produces:
      - application/json
      responses:
        "201":
          description: New food created
          schema:
            $ref: '#/definitions/models.Food'
        "500":
          description: Internal Server Error while creating a new food
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Creates a food  resource
      tags:
      - foods
  /foods/{food_id}:
    get:
      consumes:
      - application/json
      description: Retrieves  a  food with specific food id from the  orders collection
      parameters:
      - description: food_id
        in: path
        name: food_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details of a speicific food
          schema:
            $ref: '#/definitions/models.Food'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves a food with specific food id
      tags:
      - foods
    patch:
      consumes:
      - application/json
      description: Updates an existing fodd  resource in the foods collection
      parameters:
      - description: ID of the food resource to update
        in: path
        name: food_id
        required: true
        type: string
      - description: Food object
        in: body
        name: food
        required: true
        schema:
          $ref: '#/definitions/models.Food'
      produces:
      - application/json
      responses:
        "200":
          description: food got updated with new body
          schema:
            $ref: '#/definitions/models.Food'
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error while updating food
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Updates an food  resource
      tags:
      - foods
...
  /invoices:
    get:
      description: Returns an array of invoices from the invoiceCollection in restaurent
        database.
      produces:
      - application/json
      responses:
        "200":
          description: 'slice of invoices '
          schema:
            items:
              $ref: '#/definitions/models.Invoice'
            type: array
      security:
      - Security.require(true): []
      summary: Returns slice of invoices
      tags:
      - invoice
    post:
      consumes:
      - application/json
      description: Creates a invoice resource on the server
      parameters:
      - description: Invoice object
        in: body
        name: Invoice
        required: true
        schema:
          $ref: '#/definitions/models.Invoice'
      produces:
      - application/json
      responses:
        "201":
          description: New invoice created
          schema:
            $ref: '#/definitions/models.Invoice'
        "500":
          description: Internal Server Error while creating a new food
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Creates a invoice
      tags:
      - invoice
  /invoices/{invoice_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a invoice  with specific inovoice  id from the  invoices
        collection
      parameters:
      - description: invoice_id
        in: path
        name: invoice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details of a specific invoice created
          schema:
            $ref: '#/definitions/controllers.InvoiceViewFormat'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves a invoice  with specific invoice id
      tags:
      - invoice
    patch:
      consumes:
      - application/json
      description: Updates an existing invoice  resource in the invoices collection
      parameters:
      - description: ID of the invoice resource to update
        in: path
        name: invoice_id
        required: true
        type: string
      - description: Invoice object
        in: body
        name: Invoice
        required: true
        schema:
          $ref: '#/definitions/models.Invoice'
      produces:
      - application/json
      responses:
        "200":
          description: invoice got updated with new body
          schema:
            $ref: '#/definitions/models.Invoice'
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error while updating invoice
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Updates an invoice resource
      tags:
      - invoice
...
securityDefinitions:
  "":
    description: JWT token for authentication
    in: header
    name: token
    type: apiKey
swagger: "2.0"